+++ intelmq/bots/collectors/eset/collector.py
@@ -1,6 +1,7 @@
 import datetime
 import json
 from intelmq.lib.bot import CollectorBot
+from intelmq.lib.exceptions import MissingDependencyError
 
 try:
     import cabby
@@ -11,27 +12,28 @@ except ImportError:
 class ESETCollectorBot(CollectorBot):
     def init(self):
         if cabby is None:
-            raise ValueError('Library "cabby" not installed')
+            raise MissingDependencyError('cabby')
+        self.user = self.parameters.username
+        self.passwd = self.parameters.password
+        self.endpoint = self.parameters.endpoint
+        self.time_delta = int(self.parameters.time_delta)
+        self.collection = self.parameters.collection
 
     def process(self):
-        user = self.parameters.username
-        passwd = self.parameters.password
-        endpoint = self.parameters.endpoint
-        time_delta = self.parameters.time_delta
-        collection = self.parameters.collection
-
         domains = []
         end = datetime.datetime.now(datetime.timezone.utc)
-        start = end - datetime.timedelta(seconds=time_delta)  # only fetch from up to time_delta seconds ago
+        start = end - datetime.timedelta(seconds=self.time_delta)
+
+        self.logger.debug('Fetching data back to %s.', start.isoformat(timespec='seconds'))
 
         self.logger.debug('Authenticating.')
 
-        client = cabby.create_client(endpoint, discovery_path="/taxiiservice/discovery", use_https=True)
-        client.set_auth(username=user, password=passwd)
+        client = cabby.create_client(self.endpoint, discovery_path="/taxiiservice/discovery", use_https=True)
+        client.set_auth(username=self.user, password=self.passwd)
 
-        self.logger.debug('Fetching data from ESET TAXII endpoint.')
+        self.logger.debug('Authentication succeeded. Polling data from ESET TAXII endpoint.')
 
-        for item in client.poll(collection, begin_date=start, end_date=end):
+        for item in client.poll(self.collection, begin_date=start, end_date=end):
             if not item.content:
                 continue  # skip empty items
 
@@ -46,6 +48,7 @@ class ESETCollectorBot(CollectorBot):
         self.logger.debug('Submitting data.')
 
         report = self.new_report()
+        report.add("feed.url", "https://%s/taxiiservice/discovery" % self.endpoint)
         report.add('raw', json.dumps(domains))
         self.send_message(report)